Revert material restrictions on magnet holster
Currently, the material restrictions don't work well with the item it's supposed to store: the breaching shotgun. Despite the holster allowing basic `steel` material (from which the shotgun is made, along with plastic), presumably the even split between the two materials doesn't allow steel to register as a predominant material.